requirements.txt
pandas>=1.3.0
scikit-learn>=1.0.0
toml>=0.10.2
pydantic>=1.10.0
python-dotenv>=0.19.0
